Fix bundle ID and test
