+++ .github/workflows/sessions-integration-tests.yml
@@ -42,4 +42,4 @@ jobs:
             FirebaseSessions/Tests/TestApp/Shared/GoogleService-Info.plist "$plist_secret"
     - name: BuildAndTest # can be replaced with pod lib lint with CocoaPods 1.10
       run: scripts/third_party/travis/retry.sh scripts/build.sh Sessions all integration
-      
+
