+++ FirebaseSessions/Tests/TestApp/AppQualityDevApp.xcodeproj/xcshareddata/xcschemes/AppQualityDevApp_macOS.xcscheme
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1320"
-   version = "2.0">
+   LastUpgradeVersion = "1400"
+   version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
@@ -14,23 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "CFF5AB0D29004E37004A35A0"
-               BuildableName = "AppQualityDevAppWatchOS.app"
-               BlueprintName = "AppQualityDevAppWatchOS"
-               ReferencedContainer = "container:AppQualityDevApp.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "DF33DA6A28E4941400F7B88B"
-               BuildableName = "AppQualityDevApp.app"
-               BlueprintName = "AppQualityDevApp (iOS)"
+               BlueprintIdentifier = "DF33DA7028E4941400F7B88B"
+               BuildableName = "AppQualityDevApp_macOS.app"
+               BlueprintName = "AppQualityDevApp_macOS"
                ReferencedContainer = "container:AppQualityDevApp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -43,13 +29,12 @@
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
          <TestableReference
-            skipped = "NO"
-            parallelizable = "YES">
+            skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "CF94D83929070DF600308FF7"
-               BuildableName = "AppQualityDevAppUITests (iOS).xctest"
-               BlueprintName = "AppQualityDevAppUITests (iOS)"
+               BlueprintIdentifier = "DF33DA8328E4941400F7B88B"
+               BuildableName = "Tests macOS.xctest"
+               BlueprintName = "Tests macOS"
                ReferencedContainer = "container:AppQualityDevApp.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -64,15 +49,14 @@
       ignoresPersistentStateOnLaunch = "NO"
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
-      allowLocationSimulation = "YES"
-      launchAutomaticallySubstyle = "32">
+      allowLocationSimulation = "YES">
       <BuildableProductRunnable
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "CFF5AB0D29004E37004A35A0"
-            BuildableName = "AppQualityDevAppWatchOS.app"
-            BlueprintName = "AppQualityDevAppWatchOS"
+            BlueprintIdentifier = "DF33DA7028E4941400F7B88B"
+            BuildableName = "AppQualityDevApp_macOS.app"
+            BlueprintName = "AppQualityDevApp_macOS"
             ReferencedContainer = "container:AppQualityDevApp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
@@ -82,15 +66,14 @@
       shouldUseLaunchSchemeArgsEnv = "YES"
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
-      debugDocumentVersioning = "YES"
-      launchAutomaticallySubstyle = "32">
+      debugDocumentVersioning = "YES">
       <BuildableProductRunnable
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "CFF5AB0D29004E37004A35A0"
-            BuildableName = "AppQualityDevAppWatchOS.app"
-            BlueprintName = "AppQualityDevAppWatchOS"
+            BlueprintIdentifier = "DF33DA7028E4941400F7B88B"
+            BuildableName = "AppQualityDevApp_macOS.app"
+            BlueprintName = "AppQualityDevApp_macOS"
             ReferencedContainer = "container:AppQualityDevApp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
