+++ FirebaseSessions/Tests/TestApp/AppQualityDevApp.xcodeproj/xcshareddata/xcschemes/UITests_iOS.xcscheme
@@ -18,8 +18,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "CF94D83929070DF600308FF7"
-               BuildableName = "AppQualityDevAppUITests (iOS).xctest"
-               BlueprintName = "AppQualityDevAppUITests (iOS)"
+               BuildableName = "UITests_iOS.xctest"
+               BlueprintName = "UITests_iOS"
                ReferencedContainer = "container:AppQualityDevApp.xcodeproj">
             </BuildableReference>
          </TestableReference>
