--- FirebaseSessions/Tests/TestApp/AppQualityDevAppWatchOS WatchKit Extension/ContentView.swift
@@ -1,21 +1,0 @@
-//
-//  ContentView.swift
-//  AppQualityDevAppWatchOS WatchKit Extension
-//
-//  Created by Leo Zhan on 2022-10-19.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        Text("Hello, World!")
-            .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
