--- FirebaseSessions/Tests/TestApp/AppQualityDevAppWatchOS WatchKit Extension/NotificationView.swift
@@ -1,20 +1,0 @@
-//
-//  NotificationView.swift
-//  AppQualityDevAppWatchOS WatchKit Extension
-//
-//  Created by Leo Zhan on 2022-10-19.
-//
-
-import SwiftUI
-
-struct NotificationView: View {
-    var body: some View {
-        Text("Hello, World!")
-    }
-}
-
-struct NotificationView_Previews: PreviewProvider {
-    static var previews: some View {
-        NotificationView()
-    }
-}
