+++ FirebaseSessions/Tests/TestApp/Shared/AppQualityDevAppApp.swift
@@ -1,9 +1,17 @@
 //
-//  AppQualityDevAppApp.swift
-//  Shared
+// Copyright 2022 Google LLC
 //
-//  Created by Sam Edson on 9/28/22.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import SwiftUI
 import FirebaseCore
@@ -13,10 +21,8 @@ import FirebaseSessions
 struct AppQualityDevAppApp: App {
   init() {
     FirebaseApp.configure()
-    DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-//      fatalError()
-    }
   }
+
   var body: some Scene {
     WindowGroup {
       ContentView()
