--- FirebaseSessions/Tests/TestApp/Tests iOS/Tests_iOSLaunchTests.swift
@@ -1,32 +1,0 @@
-//
-//  Tests_iOSLaunchTests.swift
-//  Tests iOS
-//
-//  Created by Sam Edson on 9/28/22.
-//
-
-import XCTest
-
-class Tests_iOSLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
